enable create and delete of page

+++ src/main/java/org/knowledgeroot/app/page/PageDao.java
@@ -46,5 +46,5 @@ public interface PageDao {
      * delete page by given id
      * @param id
      */
-    void deletePageById(long id);
+    void deletePageById(Integer id);
 }
